fix: スライドのバグとコードリファクタリング
