+++ src/app/data/resultDetail.ts
@@ -1,10 +1,10 @@
-import { ResultDetailInfo } from "@/types/types";
+import { ResultDetailInfo } from '@/types/types';
 
 export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 1,
     genre: '経済の自由度',
-    genreImage: '/',
+    genreImage: '/result/type1.png',
     yourType: 'E型（経済重視派）',
     typeDescription:
       '経済重視派のあなたは、経済重視のことはもちろん、自由市場や企業の自由を支持する立場です。',
@@ -12,7 +12,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 2,
     genre: '経済の自由度',
-    genreImage: '/',
+    genreImage: '/result/type1.png',
     yourType: 'P型（進歩的）',
     typeDescription:
       '進歩的のあなたは、これはテストです。これはテストです。これはテストです。',
@@ -20,7 +20,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 3,
     genre: '政府の役割',
-    genreImage: '/',
+    genreImage: '/result/type2.png',
     yourType: 'A型（権威主義）',
     typeDescription:
       '権威主義派のあなたは、強力な政府の統制や、規制を支持し、秩序や安全を優先する立場です。',
@@ -28,7 +28,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 4,
     genre: '政府の役割',
-    genreImage: '/',
+    genreImage: '/result/type2.png',
     yourType: 'L型（自由主義）',
     typeDescription:
       '自由主義派のあなたは、これはテストです。これはテストです。これはテストです。',
@@ -36,7 +36,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 5,
     genre: '社会の保守性',
-    genreImage: '/',
+    genreImage: '/result/type3.png',
     yourType: 'L型（リベラル）',
     typeDescription:
       'リベラル派のあなたは、これはテストです。これはテストです。これはテストです。',
@@ -44,7 +44,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 6,
     genre: '社会の保守性',
-    genreImage: '/',
+    genreImage: '/result/type3.png',
     yourType: 'C型（保守的）',
     typeDescription:
       '保守派のあなたは、これはテストです。これはテストです。これはテストです。',
@@ -52,7 +52,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 7,
     genre: '国際関係',
-    genreImage: '/',
+    genreImage: '/result/type4.png',
     yourType: 'I型（国際主義）',
     typeDescription:
       '国際主義派のあなたは、これはテストです。これはテストです。これはテストです。',
@@ -60,7 +60,7 @@ export const ResultDetail: ResultDetailInfo[] = [
   {
     id: 8,
     genre: '国際関係',
-    genreImage: '/',
+    genreImage: '/result/type4  .png',
     yourType: 'N型（愛国主義）',
     typeDescription:
       '愛国主義派のあなたは、これはテストです。これはテストです。これはテストです。',
