+++ src/components/features/ScoreBar.tsx
@@ -32,10 +32,30 @@ export const ScoreBar = ({
           />
         </BarContainer>
         <ShowType>
-            {type.id === 1 && <TypeFlex><p>E</p><p>P</p></TypeFlex>}
-            {type.id === 3 && <TypeFlex><p>A</p><p>L</p></TypeFlex>}
-            {type.id === 5 && <TypeFlex><p>C</p><p>L</p></TypeFlex>}
-            {type.id === 7 && <TypeFlex><p>N</p><p>I</p></TypeFlex>}
+          {type.id === 1 && (
+            <TypeFlex>
+              <p>E</p>
+              <p>P</p>
+            </TypeFlex>
+          )}
+          {type.id === 3 && (
+            <TypeFlex>
+              <p>A</p>
+              <p>L</p>
+            </TypeFlex>
+          )}
+          {type.id === 5 && (
+            <TypeFlex>
+              <p>C</p>
+              <p>L</p>
+            </TypeFlex>
+          )}
+          {type.id === 7 && (
+            <TypeFlex>
+              <p>N</p>
+              <p>I</p>
+            </TypeFlex>
+          )}
         </ShowType>
       </ScoreBarWrapper>
     </div>
@@ -78,16 +98,16 @@ const PercentageText = styled.span`
 const ShowType = styled.div`
   margin-top: 10px;
   width: 70%;
-  margin: 10px auto 0; 
+  margin: 10px auto 0;
 `;
 
 const TypeFlex = styled.div`
   display: flex;
-  justify-content: space-between; 
-  
+  justify-content: space-between;
+
   p {
-    margin: 0; 
-    font-weight: bold; 
-    color: #333; 
+    margin: 0;
+    font-weight: bold;
+    color: #333;
   }
 `;
